feat: enable single component mode on push task
Single component mode allows only releasing of a single image from the
snapshot - the image that caused the snapshot to be created in the first
place. This is particularly useful for mono repos.

This commit adds a step to the push-snapshot-to-quay task to implement
the single component mode by reducing the snapshot to a single image if
the mode is enabled.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>